Throttle filter doesn't work for quick firing sensors
I was debugging why the throttle didn't work for me. It filtered everything, never sending any updates except for the first one. 

By looking at the code, it seems that last_input_ gets pushed forward every time a new value comes in, even if it doesn't get published. That would mean that if you have a sensor that sends data every 1s but set it to throttle to 5s, the last_input_ gets updated every 1s and thus the time difference between last_input_ and now never reaches 5s.

Remove the update to last_input_ fixes the issue. 

Maybe we could take it one step further and rename that variable to last_update_ ? Because it contains the timestamp of the last value that made it trough, not the timestamp of the last value received, which the current name implies.